requirements.txt
requests
pandas
tabulate
openpyxl
python-dotenv
